Dependencies analysis supercedes preventive tile shunting
Up until now, tiles for which products with the same date are found in the
output product directory, were ignored when searching for intersecting images.

This mecanism had a few issues when staring over (#23)
- if only VV polarized product is found, the VH one won't be produced -- and
  vice versa
- if masks are requested, they won't be produced as well.

The new dependency analysis feature (#22) is already able to ignore products
that already exist, and it doesn't have the current issues.