Deprecate `Parallelize_tiles` option
The option wasn't used anymore -- it was previously used to control downloading
of S1 products on clusters. In the new architecture this is no longer required.

Eventually we should register S1 product downloading into Dask tasks.

Note: we could have thin this option permits to run S1tiling on clusters.
Thanks to Dask, this feature is deported to the scheduler configuration.