REFACT: distinguish product from image caching/filtering
In a first step, all S1 products are found and filtered according to their
date, and their orbit.

Then, this list of all known products is filtered according to the target S2
tile to retain only the S1 products that provide enough coverage.

Eventually, the S1 products are scanned for the raster images of polarisation
compatible with the requested one(s).