ENH: Don't DL EOF when time period is already covered
This is for the case orbits are missing in the requested time period and
anywhen else, but we alreay have all the EOF files in the requested time
period.

In that case we don't even try to download anything as we already have the
related files, that we already have analysed.